Experimental change to support unique solutions. NB! Work in progress...
The current implementation creates a lot of non-unique solution puzzles. I added a check for this using the get_candidates, and testing all options. However, when this was done I realised that for the puzzles created (with only one solution) get_candidates always returns only the solution itself. Possibly get_candidates on it's own may be used to check for uniqueness (have not yet verified this). I therefore implemented a new get_pencilmarks method that properly yields hints for each cell.